extract timer action to separate class
